requirements.txt
python-docx
scikit-learn
pdfplumber
python-docx
reportlab
